unify Csr, remove CUDA 10.x ifdefs
